# [perf] introduce deep gemm group_gemm_masked as bmm