<!-- Thank you for your contribution! We appreciate it. The following guidelines will help improve your pull request and facilitate feedback. If anything is unclear, don't hesitate to submit your pull request and ask the maintainers for assistance. -->

## Motivation

per-token-group quant+deep_gemm's grouped_gemm_masked is genral faster than per-tensor quant+bmm_fp8

NOTE: Multi-batch throughput is hardly comparable because expert loads differ significantly with difference quant method.

TODO in futher PR: Optimize `_per_token_group_quant_mla_deep_gemm_masked_fp8` with CudaC

### Profile on 8xH200 DeepSeekV3 1-batch

Optimize out bmm fp8, which takes 7.9 us. The quant kenel with 2 stages takes 1.65+1.14 us and the extra zeros kernel takes 893 ns.

![image](https://github.com/user-attachments/assets/02f714e1-61be-4c56-a796-95a18e1ae291)

Deepgemm bmm kernel takes 5.7 us And the quant kernel takes 2.7 us.

![image](https://github.com/user-attachments/assets/50732d40-8bcb-4224-911b-2f542fb19976)

### Profile on 8xH200 DeepSeekV3 50-batch

Bmm fp8 takes 6.4 us. The quant kenel with 2 stages takes 2.5+1.7 us and the extra zeros kernel takes 992 ns.

![image](https://github.com/user-attachments/assets/50afb31a-e3ce-4f4a-81eb-94468bd944e8)

Deepgemm bmm kernel takes 5.7 us And the quant kernel takes 2.7 us.

![image](https://github.com/user-attachments/assets/4e3b2ee4-92ac-4444-a952-dba025ec3b34)

<!-- Explain the purpose of this PR and the goals it aims to achieve. -->

## Perf on 8xH200

About 3% improvement on decode phase

```
SGL_ENABLE_JIT_DEEPGEMM=1 SGL_ENABLE_JIT_DEEPGEMM_BMM=0 python3 -m \
sglang.bench_one_batch --model-path /DeepSeek-V3 --trust-remote-code --tp 8 \
--batch-size 1 --input-len 1000 --output-len 100

Benchmark ...
Prefill. latency: 0.17955 s, throughput:   5569.44 token/s
Decode 0.  latency: 0.01748 s, throughput:     57.20 token/s
Decode 1.  latency: 0.01744 s, throughput:     57.35 token/s
Decode 2.  latency: 0.01717 s, throughput:     58.25 token/s
Decode 3.  latency: 0.01722 s, throughput:     58.09 token/s
Decode 4.  latency: 0.01742 s, throughput:     57.40 token/s
Decode.  median latency: 0.01720 s, median throughput:     58.15 token/s
Total. latency:  1.887 s, throughput:    582.93 token/s
```

- deep gemm bmm ON
```
SGL_ENABLE_JIT_DEEPGEMM=1 SGL_ENABLE_JIT_DEEPGEMM_BMM=1 python3 -m \
sglang.bench_one_batch --model-path /DeepSeek-V3 --trust-remote-code --tp 8 \
--batch-size 1 --input-len 1000 --output-len 100

Benchmark ...
Prefill. latency: 0.16662 s, throughput:   6001.70 token/s
Decode 0.  latency: 0.01687 s, throughput:     59.27 token/s
Decode 1.  latency: 0.01663 s, throughput:     60.14 token/s
Decode 2.  latency: 0.01662 s, throughput:     60.18 token/s
Decode 3.  latency: 0.01659 s, throughput:     60.28 token/s
Decode 4.  latency: 0.01656 s, throughput:     60.38 token/s
Decode.  median latency: 0.01660 s, median throughput:     60.24 token/s
Total. latency:  1.815 s, throughput:    606.06 token/s
```

## Acc

Model: DeepSeek-V3-0304

```
SGL_ENABLE_JIT_DEEPGEMM_BMM=true python3 -m sglang.launch_server --model-path /DeepSeek-V3 --trust-remote-code --tp 8 --dp 8 --enable-dp-attention --port 31000 --disable-radix-cache
python3 benchmark/gsm8k/bench_sglang.py --num-questions 1319 --parallel 1319 --port 31000
```

- deep gemm bmm OFF 5 times: 0.935, 0.937, 0.936, 0.933, 0.939
- deep gemm bmm ON 5 times: 0.934, 0.935, 0.939, 0.934, 0.937(2-5s faster each round)

## Modifications

<!-- Describe the changes made in this PR. -->

## Checklist

- [x] Format your code according to the [Code Formatting with Pre-Commit](https://docs.sglang.ai/references/contribution_guide.html#code-formatting-with-pre-commit).
- [ ] Add unit tests as outlined in the [Running Unit Tests](https://docs.sglang.ai/references/contribution_guide.html#running-unit-tests-adding-to-ci).
- [ ] Update documentation / docstrings / example tutorials as needed, according to [Writing Documentation](https://docs.sglang.ai/references/contribution_guide.html#writing-documentation-running-docs-ci).
- [x] Provide throughput / latency benchmark results and accuracy evaluation results as needed, according to [Benchmark and Profiling](https://docs.sglang.ai/references/benchmark_and_profiling.html) and [Accuracy Results](https://docs.sglang.ai/references/accuracy_evaluation.html).
- [ ] For reviewers: If you haven't made any contributions to this PR and are only assisting with merging the main branch, please remove yourself as a co-author when merging the PR.
- [ ] Please feel free to join our Slack channel at https://slack.sglang.ai to discuss your PR.